Extract `printClassBody`
